requirements.txt
boto3==1.14.44
botocore==1.17.44
cffi==1.14.6
Cython==0.29.22
fsspec==0.9.0
newtools==2.1.322
numpy==1.20.2
pandas==1.1.1
pyarrow==2.0.0
s3fs==0.6.0
dativascrubber==1.2.150
dativatools==3.3.832
configargparse==1.5.3
